pyproject.toml
[tool.poetry]
name = "linux2mqtt"
version = "0.1.0"
description = ""
authors = ["Demian <ict1986@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
yapsy = "^1.12.2"
loguru = "^0.6.0"
sensors = {git = "https://github.com/bastienleonard/pysensors.git"}
psutil = "^5.9.4"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
